Adjusted exercise ranges - fly and one-arm row

+++ ExerciseMotionTracker/Fly.swift
@@ -10,9 +10,9 @@ import Foundation
 
 final public class Fly: ExerciseMonitor
 {
-    let EXERCISE_RANGE: Float = (Float(40) * Float(M_PI)) / Float(180)
+    let EXERCISE_RANGE: Float = (Float(32) * Float(M_PI)) / Float(180)
     var MAX_FOREARM_ANGLE: Float = (Float(86) * Float(M_PI)) / Float(180)
-    var BENT_ELBOW_CONSTRAINT: Float = (Float(20) * Float(M_PI) / Float(180))
+    var BENT_ELBOW_CONSTRAINT: Float = (Float(30) * Float(M_PI) / Float(180))
     
     // returns percent complete of a bicep curl
     override public func getPercentComplete() -> Int? {
